referral_system/requirements.txt
  django
  oauth2_provider
  django_rest_framework
  djangorestframework-simplejwt
  social_django
  social-auth-app-django
  drf_social_oauth2
  adrf
  uvicorn
